backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
opencv-python==4.8.1.78
pytesseract==0.3.10
numpy==1.24.3
Pillow==10.0.1

# EasyOCR for better text extraction
easyocr==1.7.0

# Computer Vision APIs for PII detection
google-cloud-vision==3.4.4
azure-cognitiveservices-vision-computervision==0.9.0
msrest==0.7.1

# Microsoft Presidio for advanced PII detection (fallback)
presidio-analyzer==2.2.33
presidio-anonymizer==2.2.33

# spaCy for NER (required by Presidio)
spacy==3.7.2

# Additional NLP libraries
transformers==4.35.2
torch==2.1.1 